chore: Add `useStream` hook for streaming data
The `useStream` hook has been added to handle streaming data in the client application. This hook allows for subscribing to a stream of data based on the provided job information, type, and query. It manages the state of the output data and provides a convenient way to handle different types of queries.