chore: Update build_for_prod.sh to use yarn use-prod-vault-only
The code changes rename the `build_for_prod.sh` script to `build_for_prod_vault_only.sh` and update the command `yarn use-prod` to `yarn use-prod-vault-only`. This change ensures that the script uses the appropriate production settings for the project. The update improves the clarity and accuracy of the build process.